util/requirements.txt
python_frontmatter>=1.0.0,<1.1
toml>=0.10.2,<1.0
requests>=2.28.0,<2.29
python-dotenv>=0.21.0,<0.30
tqdm>=4.60